Changes to improve PIP/windows
