+++ changedetectionio/content_fetcher.py
@@ -53,7 +53,7 @@ class Fetcher():
 
 def available_fetchers():
         import inspect
-        from changedetectionio import content_fetcher
+        import content_fetcher
         p=[]
         for name, obj in inspect.getmembers(content_fetcher):
             if inspect.isclass(obj):
